fix: re-run downstream analyses after fixing bug in comutation filter
There was a minor bug in the filter for the comutation interactions. It is documented in the lab notebook for days 2020-04-8 and 2020-04-08. The bug had no effect on LUAD and COAD comutation networks and only minor effects on those of PAAD and MM. I have not found any instances of a result or conclusion changing from this fix.